+++ sockeye/prepare_data.py
@@ -45,7 +45,7 @@ def prepare_data(args: argparse.Namespace):
     samples_per_shard = args.num_samples_per_shard
     bucketing = not args.no_bucketing
     bucket_width = args.bucket_width
-    bucket_scaling = not args.no_bucket_scaling
+    bucket_scaling = args.bucket_scaling
 
     source_paths = [args.source] + args.source_factors
     source_factor_vocab_paths = [args.source_factor_vocabs[i] if i < len(args.source_factor_vocabs)