+++ CHANGELOG.md
@@ -11,6 +11,16 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [2.1.X]
+
+### Changed
+
+- Batch sizes are now per device (ex: `--batch-size 2048 --device-ids -4` gives an effective batch size of 8192).
+- Word-based batch sizes are now a max and include padding tokens in the count.  This makes memory usage more predictable.
+- Renamed `--round-batch-sizes-to-multiple-of` to `--batch-sentences-multiple-of` for clarity.  Sizes are always rounded down.
+- Default bucketing settings changed to width 8, max sequence length 95 (96 including BOS/EOS tokens), and no bucket scaling.
+- Argument `--no-bucket-scaling` replaced with `--bucket-scaling` which is False by default.
+
 ## [2.1.7]
 
 ### Changed
@@ -112,7 +122,7 @@ Each version section may have have subsections for: _Added_, _Changed_, _Removed
 
 ### Added
 
-- Added distributed training support with Horovod/OpenMPI.  Use `horovodrun` and the `--horovod` training flag.
+- Added distributed training support with Horovod/MPI.  Use `horovodrun` and the `--horovod` training flag.
 - Added Dockerfiles that build a Sockeye image with all features enabled.  See [sockeye_contrib/docker](sockeye_contrib/docker).
 - Added `none` learning rate scheduler (use a fixed rate throughout training)
 - Added `linear-decay` learning rate scheduler